wip:Add return for Event::Setter
This implementation needs a closer look because adding a position to Event::Setter requires a lot more work to fix the resulting broken code compared to other events.

The method `register_property` can optionally push the Event::Setter event provided that the argument `register_setter_event` is `true`.
But because Event::Setter requires a position, it needs `register_property` to require a position as well as its parameter since the current parameters have no way to access position.

This introduces certain problems because even though some properties needs `register_setter_event` to be true, they don't have a clear position that can be passed to Event::Setter.
How should we handle the case when a property needs a setter event registered but has no position? For now, because of the existence of some properties with non-obvious (or perhaps absent) positions,
the position in Event::Setter will be optional i.e. `Option<SpanWithSource>`.