List possible variables, types, properties and imports when cannot find reference (#180)
* Added `get_closest`, `get_possibles_message`, `get_all_variable_names`, `get_all_named_types`, `get_property_key_names_on_a_single_type` and `get_property_as_string`  functions
* Added possibles to PropertyDoesNotExist and CannotOpenFile errors. Also added
partial_import_path to the CannotOpenFile struct as it will be needed
to find candidates for the error message. Finally, added
get_property_does_not_exist_message, a function that uses
get_possibles_mesage to create a message for the PropertyDoesNotExist
error based on the the type of the property.